Reconstruct the miner statistics output
