+++ src/common/system.cpp
@@ -8,6 +8,7 @@
 
 #include <stdlib.h>
 #include <string.h>
+#include <stdio.h>
 #include "config.h"
 
 timeMark getTimeMark()
@@ -76,4 +77,38 @@ void xsleep(unsigned seconds)
 #else
     Sleep(seconds*1000);
 #endif   
+}
+
+void printMiningStats(timeMark workBeginPoint, MineContext* mineCtx, int threadsNum, double sieveSizeInGb, unsigned blockHeight, double difficulty)
+{
+    printf(" ** block: %u, difficulty: %.3lf\n", blockHeight, difficulty);
+    
+    timeMark currentPoint = getTimeMark();    
+    uint64_t elapsedTime = usDiff(workBeginPoint, currentPoint);
+    double speed = 0.0;
+    double averageSpeed = 0.0;
+    uint64_t foundChains[MaxChainLength] = {0};
+    
+    for (int i = 0; i < threadsNum; i++) {
+        for (unsigned chIdx = 1; chIdx < MaxChainLength; chIdx++)
+            foundChains[chIdx] += mineCtx[i].foundChains[chIdx];
+        
+        double threadAvgSpeed = (sieveSizeInGb * mineCtx[i].totalRoundsNum) / (elapsedTime / 1000000.0);
+        speed += mineCtx[i].speed;
+        averageSpeed += threadAvgSpeed;
+        
+        printf("    [thread %u] %.3lfG, average: %.3lfG\n", i+1, mineCtx[i].speed, threadAvgSpeed);
+    }
+    
+    printf(" ** total speed: %.3lfG, average: %.3lfG\n", speed, averageSpeed);
+    unsigned chIdx;
+    
+    for (chIdx = 1; chIdx < MaxChainLength && foundChains[chIdx]; chIdx++) {
+        double chainsPerSec = foundChains[chIdx] / (elapsedTime / 1000000.0);
+        printf("   * chains/%u: %lu %.3lf/sec ",
+                chIdx, foundChains[chIdx], chainsPerSec);
+        if (chIdx >= 7)
+            printf("%.3lf/hour ", chainsPerSec * 3600.0);
+    }
+    printf("\n");
 }
